+++ vips/vip-552/bscmainnet.ts
@@ -1,0 +1,264 @@
+import { parseUnits } from "ethers/lib/utils";
+import { ethers } from "hardhat";
+import { NETWORK_ADDRESSES } from "src/networkAddresses";
+import { LzChainId, ProposalMeta, ProposalType } from "src/types";
+import { makeProposal } from "src/utils";
+
+const { bscmainnet, ethereum, unichainmainnet, zksyncmainnet, arbitrumone } = NETWORK_ADDRESSES;
+
+export const ADAPTER_PARAMS = ethers.utils.solidityPack(["uint16", "uint256"], [1, 300000]);
+export const BRIDGE_FEES_BSC = parseUnits("0.5", 18);
+
+export const ZKSYNC_XVS_BRIDGE_AMOUNT = parseUnits("126", 18);
+export const ZKSYNC_XVS_STORE = "0x84266F552756cBed893b1FFA85248cD99501e3ce";
+export const ZKSYNC_SPEED = parseUnits("0.00011632", 18); // 10.05 XVS/day
+
+export const UNICHAIN_XVS_BRIDGE_AMOUNT = parseUnits("205", 18);
+export const UNICHAIN_XVS_STORE = "0x0ee4b35c2cEAb19856Bf35505F81608d12B2a7Bb";
+export const UNICHAIN_SPEED = parseUnits("0.00003588", 18); // 3.1 XVS/day
+
+export const BSC_XVS_BRIDGE = "0xf8F46791E3dB29a029Ec6c9d946226f3c613e854";
+export const BSC_XVS_VAULT_TREASURY = "0x269ff7818DB317f60E386D2be0B259e1a324a40a";
+export const BSC_XVS_STORE = "0x1e25CF968f12850003Db17E0Dba32108509C4359";
+export const BSC_XVS_AMOUNT = parseUnits("99174", 18);
+export const BSC_SPEED = parseUnits("0.01213542", 18); // 1,398 XVS/day
+
+export const ETH_XVS_VAULT_TREASURY = "0xaE39C38AF957338b3cEE2b3E5d825ea88df02EfE";
+export const ETH_XVS_STORE = "0x1Db646E1Ab05571AF99e47e8F909801e5C99d37B";
+export const ETH_XVS_AMOUNT = parseUnits("2129", 18);
+export const ETH_SPEED = parseUnits("0.00378889", 18); // 27.28 XVS/day
+
+export const ARB_XVS_VAULT_TREASURY = "0xb076D4f15c08D7A7B89466327Ba71bc7e1311b58";
+export const ARB_XVS_STORE = "0x507D9923c954AAD8eC530ed8Dedb75bFc893Ec5e";
+export const ARB_XVS_AMOUNT = parseUnits("385", 18);
+export const ARB_SPEED = parseUnits("0.00002593", 18); // 2.24 XVS/day
+
+export const vip552 = async () => {
+  const meta: ProposalMeta = {
+    version: "v2",
+    title: "VIP-552 Quarterly XVS Buyback and Funds Allocation",
+    description: `## Summary
+
+This VIP outlines the protocol’s Quarterly Buyback and Funds Allocation strategy, in accordance with [Tokenomics V4.1](https://docs-v4.venus.io/governance/tokenomics), based on fees generated in Q3 2025 (details below).
+
+The XVS Buyback was executed gradually over Q3 2025. This VIP will redistribute these buybacks to XVS stakers by transferring the buybacks to the XVS Stores and adjusting the reward speeds on each chain.
+
+If approved, this VIP will:
+
+- Transfer XVS from the XVS Vault Treasuries to the XVS Stores on BNB, Ethereum and Arbitrum, enabling distribution of XVS Vault rewards throughout the quarter.
+- Bridge XVS to Unichain and zkSync to fund the XVS Vault rewards proportionally to the revenue generated on these chains.
+- Update reward speeds in line with the XVS bought back during Q3 2025.
+    - On BNB chain: Rewards will be topped up with 308.7 XVS/day legacy allocation.
+    - On other chains: Rewards will be based solely on revenue generated, with no additional allocation (see [XVS Vault Emissions Adjustment Proposal](https://community.venus.io/t/xvs-vault-emissions-adjustment/5405)).
+
+## Q3 2025 Fees and allocation
+
+### **BNB Chain**
+
+**Protocol Fees**
+
+- Reserves: $1,089,607
+- Liquidation Fees: $2,500,149 (*$2,036,605 generated by the [liquidation of the account 0x7fd8…202a](https://bscscan.com/tx/0xee9928b8d1a212f4d7b7e9dca97598394005a7b8fef56856e52351bc7921be43))*
+- Total: $3,589,756
+
+**Tokenomics Allocations**
+
+- XVS Buybacks: $717,951 (99,174 XVS converted)
+- Venus Prime: $217,921
+- Treasury: $1,038,493
+- Risk Fund: $717,951
+- BNB Burn: $897,439
+
+### **Ethereum**
+
+**Protocol Fees**
+
+- Reserves: $12,206
+- Liquidation Fees: $5,234
+- Total: $17,840
+
+**Tokenomics Allocations**
+
+- XVS Buybacks: $3,568 (2,129 XVS converted)
+- Venus Prime: $2,521
+- Treasury: $3,814
+
+### **Arbitrum one**
+
+**Protocol Fees**
+
+- Reserves: $5,890
+- Liquidation Fees: $350
+- Total: $6,240
+
+**Tokenomics Allocations**
+
+- XVS Buybacks: $1,248 (385 XVS converted)
+- Venus Prime: $1,178
+- Treasury: $3,814
+
+### **ZKsync Era**
+
+**Protocol Fees**
+
+- Reserves: $4,113
+- Liquidation Fees: $107
+- Total: $4,220
+
+**Tokenomics Allocations**
+
+- XVS Buybacks: $844 (126 XVS)
+- Venus Prime: $823
+- Treasury: $2,553
+
+### Unichain
+
+**Protocol Fees**
+
+- Reserves: $6,850
+- Liquidation Fees: $1
+- Total: $6,851
+
+**Tokenomics Allocations**
+
+- XVS Buybacks: $1,370 (205 XVS)
+- Venus Prime: $1,370
+- Treasury: $4,111
+
+## Q4 2025 XVS Vault Rewards
+
+The XVS available for Vault Rewards will come from:
+
+1. **XVS Buybacks** accumulated throughout Q3 in the XVS Vault Treasuries.
+2. Any **spare XVS** remaining in the XVS Stores (assuming 9-Oct-25 execution date).
+3. A **fixed Base Reward allocation** of 308.7 XVS/day on **BNB chain only.**
+
+Any Buyback allocation in the XVS Converters that has not yet been converted to XVS will not be taken into account in the Rewards distribution this quarter.
+
+**If approved, this VIP will perform the following actions:**
+
+**BNB Chain:**
+
+- **XVS Buyback Allocation**: Transfer **99,174 XVS** from the [XVS Vault Treasury](https://bscscan.com/address/0x269ff7818DB317f60E386D2be0B259e1a324a40a) to the [XVS Store](https://bscscan.com/address/0x1e25CF968f12850003Db17E0Dba32108509C4359).
+- **XVS Vault Speed:** Set the new XVS Vault reward speed to  **1,398 XVS/day** (includes Q3 buybacks + 308.7 XVS/day Base Rewards)
+
+**Ethereum:**
+
+- **XVS Buyback Allocation**: Transfer **2,129 XVS** from the [XVS Vault Treasury](https://etherscan.io/address/0xaE39C38AF957338b3cEE2b3E5d825ea88df02EfE) to the [XVS Store](https://etherscan.io/address/0x1Db646E1Ab05571AF99e47e8F909801e5C99d37B).
+- **XVS Vault Speed**: Set the new XVS Vault reward speed to **27.28 XVS/day** (includes Q3 buybacks + 354 XVS leftover in the [XVS Store](https://etherscan.io/address/0x1Db646E1Ab05571AF99e47e8F909801e5C99d37B))
+
+**Arbitrum one:**
+
+- **XVS Buyback Allocation:** Transfer **385 XVS** from the [XVS Vault Treasury](https://arbiscan.io/address/0xb076D4f15c08D7A7B89466327Ba71bc7e1311b58) to the [XVS Store](https://arbiscan.io/address/0x507D9923c954AAD8eC530ed8Dedb75bFc893Ec5e).
+- **XVS Vault Speed:** Set the new XVS Vault reward speed to **2.24 XVS/day** (includes Q3 buybacks - 181 XVS to cover slight deficit in Q3)
+
+**ZKsync Era:**
+
+- **XVS Buyback Allocation**: Bridge **126 XVS** from the [Core Pool Comptroller](https://bscscan.com/address/0xfD36E2c2a6789Db23113685031d7F16329158384) to the [XVS Store](https://explorer.zksync.io/address/0x84266F552756cBed893b1FFA85248cD99501e3ce). The XVS Converter and the XVS Vault Treasury have not yet been deployed on this chain so the buybacks don’t take place automatically.
+- **XVS Vault Speed:** Set the new XVS Vault reward speed to **10.05 XVS/day** (includes Q3 buybacks + 789 XVS leftover in the [XVS Store](https://explorer.zksync.io/address/0x84266F552756cBed893b1FFA85248cD99501e3ce))
+
+**Unichain:**
+
+- **XVS Buyback Allocation**: Bridge **205 XVS** from [Core Pool Comptroller](https://bscscan.com/address/0xfD36E2c2a6789Db23113685031d7F16329158384) to the [XVS Store](https://arbiscan.io/address/0x507D9923c954AAD8eC530ed8Dedb75bFc893Ec5e). The XVS Converter and the XVS Vault Treasury have not yet been deployed on this chain so the buybacks don’t take place automatically.
+- **XVS Vault Speed**: Set the new XVS Vault reward speed to **3.1 XVS/day** (includes Q3 buybacks + 79 XVS leftover in the [XVS Store](https://arbiscan.io/address/0x507D9923c954AAD8eC530ed8Dedb75bFc893Ec5e))**.**
+
+### **References**
+
+- [VIP simulation](https://github.com/VenusProtocol/vips/pull/619)
+- [Tokenomics V4.1](https://docs-v4.venus.io/governance/tokenomics)
+- [XVS Vault Emissions Adjustment](https://community.venus.io/t/xvs-vault-emissions-adjustment/5405)
+- [VIP-529 Quarterly XVS Buyback and Funds Allocation](https://app.venus.io/#/governance/proposal/529?chainId=56)`,
+    forDescription: "I agree that Venus Protocol should proceed with this proposal",
+    againstDescription: "I do not think that Venus Protocol should proceed with this proposal",
+    abstainDescription: "I am indifferent to whether Venus Protocol proceeds or not",
+  };
+
+  return makeProposal(
+    [
+      {
+        target: bscmainnet.UNITROLLER,
+        signature: "_grantXVS(address,uint256)",
+        params: [bscmainnet.NORMAL_TIMELOCK, ZKSYNC_XVS_BRIDGE_AMOUNT.add(UNICHAIN_XVS_BRIDGE_AMOUNT)],
+      },
+      {
+        target: bscmainnet.XVS,
+        signature: "approve(address,uint256)",
+        params: [BSC_XVS_BRIDGE, ZKSYNC_XVS_BRIDGE_AMOUNT.add(UNICHAIN_XVS_BRIDGE_AMOUNT)],
+      },
+      {
+        target: BSC_XVS_BRIDGE,
+        signature: "sendFrom(address,uint16,bytes32,uint256,(address,address,bytes))",
+        params: [
+          bscmainnet.NORMAL_TIMELOCK,
+          LzChainId.zksyncmainnet,
+          ethers.utils.defaultAbiCoder.encode(["address"], [ZKSYNC_XVS_STORE]),
+          ZKSYNC_XVS_BRIDGE_AMOUNT,
+          [bscmainnet.NORMAL_TIMELOCK, ethers.constants.AddressZero, ADAPTER_PARAMS],
+        ],
+        value: BRIDGE_FEES_BSC.toString(),
+      },
+      {
+        target: BSC_XVS_BRIDGE,
+        signature: "sendFrom(address,uint16,bytes32,uint256,(address,address,bytes))",
+        params: [
+          bscmainnet.NORMAL_TIMELOCK,
+          LzChainId.unichainmainnet,
+          ethers.utils.defaultAbiCoder.encode(["address"], [UNICHAIN_XVS_STORE]),
+          UNICHAIN_XVS_BRIDGE_AMOUNT,
+          [bscmainnet.NORMAL_TIMELOCK, ethers.constants.AddressZero, ADAPTER_PARAMS],
+        ],
+        value: BRIDGE_FEES_BSC.toString(),
+      },
+      {
+        target: zksyncmainnet.XVS_VAULT_PROXY,
+        signature: "setRewardAmountPerBlockOrSecond(address,uint256)",
+        params: [zksyncmainnet.XVS, ZKSYNC_SPEED],
+        dstChainId: LzChainId.zksyncmainnet,
+      },
+      {
+        target: unichainmainnet.XVS_VAULT_PROXY,
+        signature: "setRewardAmountPerBlockOrSecond(address,uint256)",
+        params: [unichainmainnet.XVS, UNICHAIN_SPEED],
+        dstChainId: LzChainId.unichainmainnet,
+      },
+      {
+        target: BSC_XVS_VAULT_TREASURY,
+        signature: "fundXVSVault(uint256)",
+        params: [BSC_XVS_AMOUNT],
+      },
+      {
+        target: bscmainnet.XVS_VAULT_PROXY,
+        signature: "setRewardAmountPerBlockOrSecond(address,uint256)",
+        params: [bscmainnet.XVS, BSC_SPEED],
+      },
+      {
+        target: ETH_XVS_VAULT_TREASURY,
+        signature: "fundXVSVault(uint256)",
+        params: [ETH_XVS_AMOUNT],
+        dstChainId: LzChainId.ethereum,
+      },
+      {
+        target: ethereum.XVS_VAULT_PROXY,
+        signature: "setRewardAmountPerBlockOrSecond(address,uint256)",
+        params: [ethereum.XVS, ETH_SPEED],
+        dstChainId: LzChainId.ethereum,
+      },
+      {
+        target: ARB_XVS_VAULT_TREASURY,
+        signature: "fundXVSVault(uint256)",
+        params: [ARB_XVS_AMOUNT],
+        dstChainId: LzChainId.arbitrumone,
+      },
+      {
+        target: arbitrumone.XVS_VAULT_PROXY,
+        signature: "setRewardAmountPerBlockOrSecond(address,uint256)",
+        params: [arbitrumone.XVS, ARB_SPEED],
+        dstChainId: LzChainId.arbitrumone,
+      },
+    ],
+    meta,
+    ProposalType.REGULAR,
+  );
+};
+
+export default vip552;
